seperate get bill types function to external file relates #35
